pyproject.toml
[tool.poetry]
name = "spicefit"
version = "0.1.0"
description = ""
authors = ["adolliou <antoine.dolliou@universite-paris-saclay.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
numpy = "^2.1.2"
astropy = ">=6.1.4"
sospice = ">=0.0.8"
pyyaml = ">=6.0.2"
sunpy = "^6.0.4"
matplotlib = "=3.9.0"
mpl-animators = ">=1.0.0"
reproject = ">=0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
